pyproject.toml
[project]
name = "orange-checker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy==2.2.6",
    "openai==2.8.1",
    "pillow==12.0.0",
    "pyqt6==6.10.0",
    "requests==2.32.5",
    "torch",
    "torchvision",
]
[tool.uv]
add-bounds = "exact"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver2"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "sys_platform != 'win32'" }]
torchvision = [{ index = "pytorch-cpu", marker = "sys_platform != 'win32'" }]

[dependency-groups]
dev = ["commitizen==4.10.0", "pre-commit==4.4.0", "ruff==0.14.5"]
